feature(server): Create a base server - separate server file from entry point of the app - put middlewares on the entry point file - server on other point [finishes #169642557]
